// src/interfaces/components.ts
export type TextInputSize = "xs" | "sm" | "md" | "lg" | "xl";
export type TextInputVariant =
  | "primary"
  | "secondary"
  | "ghost"
  | "danger"
  | "outline";
export type TextInputShape = "rounded" | "square" | "circle";

export type ButtonSize = "xs" | "sm" | "md" | "lg" | "xl";
export type ButtonVariant =
  | "primary"
  | "secondary"
  | "danger"
  | "ghost"
  | "outline";
export type ButtonShape = "rounded" | "square" | "circle";

export type IconButtonSize = ButtonSize;
export type IconButtonVariant = ButtonVariant;
export type IconButtonShape = ButtonShape;